added uploadPic and uploadResume to helpers
